+++ docs/conf.py
@@ -51,7 +51,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'Aldryn Boilerplate Standard'
+project = u'Aldryn Boilerplate Bootstrap 3'
 copyright = u'2014, Divio AG'
 
 # The version info for the project you're documenting, acts as replacement for