+++ static/js/addons/cl.utils.js
@@ -84,6 +84,7 @@ var Cl = window.Cl || {};
         setStorage: function (token, value) {
             if (typeof (Storage) !== void(0)) {
                 localStorage.setItem(token, value);
+                return value;
             }
         },