+++ tests/test.cl.utils.js
@@ -10,7 +10,7 @@
 
     describe('Cl.Utils', function () {
         it('exists', function () {
-            expect(Cl.Utils).toEqual(jasmine.any(Object));
+            expect(Cl.Utils).toBeDefined();
         });
 
         describe('._document()', function () {
@@ -26,6 +26,47 @@
                 spyOn(Cl.Utils, '_consoleWrapper');
                 Cl.Utils._document();
                 expect(Cl.Utils._consoleWrapper).toHaveBeenCalled();
+            });
+        });
+
+        describe('.redirectTo()', function () {
+            it('forwards to a new url', function () {
+                expect(window.location.href).not.toMatch('#testRedirect');
+                Cl.Utils.redirectTo('#testRedirect');
+                expect(window.location.href).toMatch('#testRedirect');
+            });
+        });
+
+        describe('.setStorage() and .getStorage()', function () {
+            beforeEach(function () {
+                localStorage.clear();
+            });
+
+            it('use the local storage of the browser', function () {
+                if (typeof (Storage) !== void(0)) {
+                    expect(Storage).toEqual(jasmine.any(Object));
+                } else {
+                    expect(Storage).toThrowError(ReferenceError);
+                }
+            });
+
+            it('one can store a value', function () {
+                expect(localStorage.getItem('test#1')).toBeNull();
+                expect(Cl.Utils.setStorage).not.toThrow();
+
+                var returnValue = Cl.Utils.setStorage('test#1', 'true');
+                expect(returnValue).toBe('true');
+                expect(localStorage.getItem('test#1')).toBe('true');
+            });
+
+            it('the other retrieve a value', function () {
+                expect(localStorage.getItem('test#2')).toBeNull();
+                expect(Cl.Utils.getStorage).not.toThrow();
+
+                Cl.Utils.setStorage('test#2', 'true');
+                var returnValue = Cl.Utils.getStorage('test#2');
+                expect(returnValue).toBe('true');
+                expect(localStorage.getItem('test#2')).toBe('true');
             });
         });
     });